Flowsheet Separation fix
