+++ src/pages/flowsheet/FlowsheetPage.js
@@ -1,695 +1,22 @@
 import {
-    AspectRatio,
-    Avatar,
-    Box,
-    Checkbox,
-    Tooltip,
-    Chip,
-    CircularProgress,
-    Divider,
-    FormControl,
-    FormLabel,
-    IconButton,
-    Input,
-    Sheet,
-    Stack,
-    Typography,
-    Button,
-  } from "@mui/joy";
-  import DragIndicatorIcon from "@mui/icons-material/DragIndicator";
-  import PlayArrowIcon from '@mui/icons-material/PlayArrow';
-  import ClearIcon from '@mui/icons-material/Clear';
-  import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
-  import PortableWifiOffIcon from '@mui/icons-material/PortableWifiOff';
-  import WifiTetheringIcon from '@mui/icons-material/WifiTethering';
-  import TroubleshootIcon from "@mui/icons-material/Troubleshoot";
-  import MicIcon from "@mui/icons-material/Mic";
-  import HeadphonesIcon from "@mui/icons-material/Headphones";
-  import TimerIcon from "@mui/icons-material/Timer";
-  import LogoutIcon from "@mui/icons-material/Logout";
-  import React, { useCallback, useEffect, useRef, useState } from "react";
-  import { getArtwork } from "../../services/artwork/artwork-service";
-  import { ArtistAvatar } from "../../components/catalog/ArtistAvatar";
-  import { RotationAvatar } from "../../components/flowsheet/RotationAvatar";
+  Divider,
+  Sheet,
+  Stack
+} from "@mui/joy";
+import React from "react";
+import AddToFlowsheetSearch from "../../components/flowsheet/AddToFlowsheetSearch";
+import FlowsheetEntry from "../../components/flowsheet/FlowsheetEntry";
 import { useFlowsheet } from "../../services/flowsheet/flowsheet-context";
-import { useAuth } from "../../services/authentication/authentication-context";
-import { useLive } from "../../services/flowsheet/live-context";
   
   const FlowSheetPage = () => {
 
-    const [searchResults, setSearchResults] = useState([]); 
-    
-    const { user } = useAuth();
-    const { queue, entries, addToEntries } = useFlowsheet();
+    const { queue, entries } = useFlowsheet();
 
-    const { live, setLive } = useLive();
-
-    const switchLive = () => {
-        if (live) {
-            setLive(false);
-            addToEntries({
-                message: `DJ ${user.djName} left at ${new Date().toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' })}`,
-            });
-        } else {
-            setLive(true);
-            addToEntries({
-                message: `DJ ${user.djName} joined at ${new Date().toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' })}`,
-            });
-        }
-    }
-
-    const addTalkset = () => {
-        addToEntries({
-            message: "Talkset",
-        });
-    }
-
-    const addBreakpoint = () => {
-        const nextHour = new Date();
-        nextHour.setHours(nextHour.getHours() + 1);
-        nextHour.setMinutes(0);
-        nextHour.setSeconds(0);
-        addToEntries({
-            message: `${nextHour.toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' })} Breakpoint`,
-        });
-    }
-
-    // THIS IS WHERE THE PAGE BEGINS ---------------------------------------------
-    const searchRef = useRef(null);
-    const [searching, setSearching] = useState(false);
-  
-    const [searchstring, setSearchstring] = useState("");
-    const [selected, setSelected] = useState(0);
-    const [searchType, setSearchType] = useState("title"); // ['title', 'artist', 'album', 'label']
-    const [fieldStrings, setFieldStrings] = useState({
-      title: "",
-      artist: "",
-      album: "",
-      label: "",
-    });
-  
-    const handleSearchDown = (e) => {
-    if (!live) return;
-      if (e.key === "/") {
-        e.preventDefault();
-        // get input child from searchRef
-        const input = searchRef.current.querySelector("input");
-        input.focus();
-      }
-      if (searchRef.current.querySelector("input") === document.activeElement) {
-        if (e.key === "Escape") {
-          e.preventDefault();
-          e.target.blur();
-        } else if (e.key === "Tab") {
-          e.preventDefault();
-          let newSearchMap = null;
-          if (e.shiftKey) {
-            newSearchMap = {
-              title: "label",
-              artist: "title",
-              album: "artist",
-              label: "album",
-            };
-          } else {
-            newSearchMap = {
-              title: "artist",
-              artist: "album",
-              album: "label",
-              label: "title",
-            };
-          }
-          setSearchstring("");
-          setSearchType((previous) => newSearchMap[previous]);
-        } else if (e.keyCode === 38) {
-          e.preventDefault();
-          setSelected((previous) => Math.max(0, previous - 1));
-        } else if (e.keyCode === 40) {
-          e.preventDefault();
-          setSelected((previous) =>
-            Math.min(searchResults.length - 1, previous + 1)
-          );
-        } else if (e.key === "Enter") {
-          e.preventDefault();
-          console.log("Selected " + selected);
-          closeSearch();
-        }
-      }
-    };
-  
-    const closeSearch = () => {
-      setSearching(false);
-      setSearchstring("");
-      setSelected(0);
-      setSearchType("title");
-      setFieldStrings({
-        title: "",
-        artist: "",
-        album: "",
-        label: "",
-      });
-    };
-  
-    const handleSearchFocused = (e) => {
-      setSearching(true);
-    };
-  
-    const handleSearchChange = (e) => {
-      setSearchstring(e.target.value);
-      let newFieldStrings = { ...fieldStrings };
-      newFieldStrings[searchType] = e.target.value;
-      setFieldStrings(newFieldStrings);
-    };
-  
-    useEffect(() => {
-    document.removeEventListener("keydown", handleSearchDown);
-      document.addEventListener("keydown", handleSearchDown);
-      return () => {
-        document.removeEventListener("keydown", handleSearchDown);
-      };
-    }, [live]);
-  
     // THIS IS WHERE THE PAGE RENDER BEGINS ---------------------------------------------
     return (
       <div>
       {/* HEADER AREA */}
-        <Box
-          sx={{
-            display: "flex",
-            alignItems: "center",
-            my: 1,
-            gap: 1,
-            flexWrap: "wrap",
-            "& > *": {
-              minWidth: "clamp(0px, (500px - 100%) * 999, 100%)",
-              flexGrow: 1,
-            },
-          }}
-        >
-          <Typography level="h1">Flowsheet</Typography>
-          <Box sx={{ flex: 999 }}></Box>
-          <Button
-            variant={(live) ? "solid" : "outlined"}
-            color={(live) ? "primary" : "neutral"}
-            startDecorator={(live) ? <WifiTetheringIcon /> : <PortableWifiOffIcon />}
-            onClick={switchLive}
-          >
-            {live ? "You are Live" : "You are Offline"}
-          </Button>
-        </Box>
-        {/* SEARCH AREA */}
-        <Stack direction="row" spacing={1}>
-          <FormControl size="sm" sx={{ flex: 1 }}>
-            {searching && (
-              <Sheet
-                variant="outlined"
-                sx={{
-                  minHeight: "60px",
-                  position: "absolute",
-                  top: -5,
-                  left: -5,
-                  right: -5,
-                  zIndex: 1,
-                  borderRadius: "md",
-                  transition: "height 0.2s ease-in-out",
-                  boxShadow: "0px 34px 24px -9px rgba(0,0,0,0.7)",
-                }}
-              >
-                <Box
-                  sx={{
-                    mt: "40px",
-                    position: "relative",
-                    minHeight: "40px",
-                    pb: "40px",
-                    transition: "height 0.2s ease-in-out",
-                  }}
-                >
-                  {searchstring.length +
-                    Object.keys(fieldStrings).reduce((total, item) => {
-                      return total + fieldStrings[item].length;
-                    }, 0) >
-                    0 && (
-                    <Box
-                      sx={{
-                        p: 1,
-                        backgroundColor:
-                          selected == 0 ? "primary.700" : "transparent",
-                        cursor: "pointer",
-                      }}
-                      onMouseOver={() => setSelected(0)}
-                    >
-                      <Typography level="body4">
-                        CREATE A NEW ENTRY WITH THE FOLLOWING FIELDS:
-                      </Typography>
-                      <Stack direction="row" justifyContent="space-between">
-                        {Object.keys(fieldStrings).map((item, index) =>
-                          fieldStrings[item].length > 0 ? (
-                            <Chip key={item}>
-                              <Typography level="body2">
-                                {item}: {fieldStrings[item]}
-                              </Typography>
-                            </Chip>
-                          ) : (
-                            <div></div>
-                          )
-                        )}
-                      </Stack>
-                    </Box>
-                  )}
-                  <>
-                    <Divider />
-                    <Box
-                      sx={{
-                        p: 1,
-                      }}
-                    >
-                      <Typography level="body4">FROM YOUR MAIL BIN</Typography>
-                    </Box>
-                    <Stack direction="column">
-                      <Stack
-                        direction="row"
-                        justifyContent="space-between"
-                        sx={{
-                          p: 1,
-                          backgroundColor:
-                            selected == 1 ? "primary.700" : "transparent",
-                          cursor: "pointer",
-                        }}
-                        onMouseOver={() => setSelected(1)}
-                      >
-                        <ArtistAvatar
-                          artist={{
-                            genre: "Rock",
-                            lettercode: "AB",
-                            numbercode: "128",
-                            entry: "1",
-                          }}
-                        />
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            SONG
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            {fieldStrings["title"]}
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            ALBUM
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Album Name
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            ARTIST
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Artist Name
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            LABEL
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Label Name
-                          </Typography>
-                        </Stack>
-                      </Stack>
-                    </Stack>
-                  </>
-                  <>
-                    <Divider />
-                    <Box
-                      sx={{
-                        p: 1,
-                      }}
-                    >
-                      <Typography level="body4">ROTATION</Typography>
-                    </Box>
-                    <Stack direction="column">
-                      <Stack
-                        direction="row"
-                        justifyContent="space-between"
-                        sx={{
-                          p: 1,
-                          backgroundColor:
-                            selected == 2 ? "primary.700" : "transparent",
-                          cursor: "pointer",
-                        }}
-                        onMouseOver={() => setSelected(2)}
-                      >
-                        <RotationAvatar rotation="M" />
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            SONG
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            {fieldStrings["title"]}
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            ALBUM
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Album Name
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            ARTIST
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Artist Name
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            LABEL
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Label Name
-                          </Typography>
-                        </Stack>
-                      </Stack>
-                    </Stack>
-  
-                    <Stack direction="column">
-                      <Stack
-                        direction="row"
-                        justifyContent="space-between"
-                        sx={{
-                          p: 1,
-                          backgroundColor:
-                            selected == 3 ? "primary.700" : "transparent",
-                          cursor: "pointer",
-                        }}
-                        onMouseOver={() => setSelected(3)}
-                      >
-                        <RotationAvatar rotation="H" />
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            SONG
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            {fieldStrings["title"]}
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            ALBUM
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Album Name
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            ARTIST
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Artist Name
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            LABEL
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Label Name
-                          </Typography>
-                        </Stack>
-                      </Stack>
-                    </Stack>
-                  </>
-                  <>
-                    <Divider />
-                    <Box
-                      sx={{
-                        p: 1,
-                      }}
-                    >
-                      <Typography level="body4">CATALOG</Typography>
-                    </Box>
-                    <Stack direction="column">
-                      <Stack
-                        direction="row"
-                        justifyContent="space-between"
-                        sx={{
-                          p: 1,
-                          backgroundColor:
-                            selected == 4 ? "primary.700" : "transparent",
-                          cursor: "pointer",
-                        }}
-                        onMouseOver={() => setSelected(4)}
-                      >
-                        <ArtistAvatar
-                          artist={{
-                            genre: "Hiphop",
-                            lettercode: "AB",
-                            numbercode: "128",
-                            entry: "1",
-                          }}
-                        />
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            SONG
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            {fieldStrings["title"]}
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            ALBUM
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Album Name
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            ARTIST
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Artist Name
-                          </Typography>
-                        </Stack>
-                        <Stack direction="column" sx={{ width: "calc(20%)" }}>
-                          <Typography level="body4" sx={{ mb: -1 }}>
-                            LABEL
-                          </Typography>
-                          <Typography
-                            sx={{
-                              whiteSpace: "nowrap",
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                            }}
-                          >
-                            Label Name
-                          </Typography>
-                        </Stack>
-                      </Stack>
-                    </Stack>
-                  </>
-                  <Divider />
-                  <Stack
-                    direction="row"
-                    justifyContent="flex-end"
-                    alignItems="center"
-                    spacing={0.25}
-                    sx={{
-                      position: "absolute",
-                      bottom: 0,
-                      left: 0,
-                      right: 0,
-                      height: "40px",
-                      p: 1,
-                      "& > *": {
-                        lineHeight: "0.5rem !important",
-                      },
-                    }}
-                  >
-                    <Chip variant="soft" size="sm" color="neutral">
-                      <Typography level="body5">TAB</Typography>
-                    </Chip>
-                    <Typography level="body4">switches search fields</Typography>
-                    <Chip variant="soft" size="sm" color="neutral">
-                      <Typography level="body5">SHIFT + TAB</Typography>
-                    </Chip>
-                    <Typography level="body4">goes back a field</Typography>
-                    <Chip variant="soft" size="sm" color="neutral">
-                      <Typography level="body5">UP ARROW</Typography>
-                    </Chip>
-                    <Typography level="body4">
-                      selects the previous entry
-                    </Typography>
-                    <Chip variant="soft" size="sm" color="neutral">
-                      <Typography level="body5">DOWN ARROW</Typography>
-                    </Chip>
-                    <Typography level="body4">selects the next entry</Typography>
-                    <Chip variant="soft" size="sm" color="neutral">
-                      <Typography level="body5">ENTER</Typography>
-                    </Chip>
-                    <Typography level="body4">
-                      adds the result to the queue
-                    </Typography>
-                    <Chip variant="soft" size="sm" color="neutral">
-                      <Typography level="body5">SHIFT + ENTER</Typography>
-                    </Chip>
-                    <Typography level="body4">
-                      sets the current result playing
-                    </Typography>
-                  </Stack>
-                </Box>
-              </Sheet>
-            )}
-            <Input
-                disabled={!live}
-              ref={searchRef}
-              placeholder={
-                searching
-                  ? (searchType != "title") ? `Enter ${searchType}` : "Start by providing a song title"
-                  : "Press  /  to search or start typing"
-              }
-              startDecorator={<TroubleshootIcon />}
-              endDecorator={
-                <>
-                <IconButton
-                  variant="outlined"
-                  color="neutral"
-                  onClick={() => {
-                    const input = searchRef.current.querySelector("input");
-                    input.value = "";
-                    input.focus();
-                  }}
-                >
-                  /
-                </IconButton>
-                </>
-              }
-              onFocus={handleSearchFocused}
-              onBlur={closeSearch}
-              value={searchstring}
-              onChange={handleSearchChange}
-              sx={{
-                zIndex: 2,
-              }}
-            />
-          </FormControl>
-          <Tooltip
-            placement="top"
-            size="sm"
-            title="Add a Breakpoint"
-            variant="outlined"
-          >
-            <IconButton size="sm" variant="solid" color="warning" onClick={addBreakpoint} disabled={!live}>
-              <TimerIcon />
-            </IconButton>
-          </Tooltip>
-          <Tooltip
-            placement="top"
-            size="sm"
-            title="Add a Talkset"
-            variant="outlined"
-          >
-            <IconButton size="sm" variant="solid" color="success" onClick={addTalkset} disabled={!live}>
-              <MicIcon />
-            </IconButton>
-          </Tooltip>
-        </Stack>
+      <AddToFlowsheetSearch />
         {/* FLOWSHEET AREA */}
         <Sheet
           sx={{
@@ -736,339 +63,6 @@ import { useLive } from "../../services/flowsheet/live-context";
         </Sheet>
       </div>
     );
-  };
-  
-  const FlowsheetEntry = (props) => {
-    const [image, setImage] = useState(null);
-
-    const [canClose, setCanClose] = useState(false);
-  
-    const getImage = useCallback(async () => {
-      let storedArtwork = sessionStorage.getItem(
-        `${props.releaseAlbum}-${props.releaseArtist}`
-      );
-      if (storedArtwork) return storedArtwork;
-      try {
-        let retrievedArtwork = await getArtwork({
-          title: props.releaseAlbum,
-          artist: props.releaseArtist,
-        });
-        // THE CONVENTION IS ALBUM THEN ARTIST IN THIS APP
-        sessionStorage.setItem(
-          `${props.releaseAlbum}-${props.releaseArtist}`,
-          retrievedArtwork
-        );
-        return retrievedArtwork;
-      } catch (e) {
-        sessionStorage.setItem(
-          `${props.releaseAlbum}-${props.releaseArtist}`,
-          ""
-        );
-        return "";
-      }
-    }, [props.releaseAlbum, props.releaseArtist]);
-  
-    useEffect(() => {
-      getImage().then((image) => {
-        setImage(image);
-      });
-    }, [getImage]);
-  
-    switch (props.type) {
-      case "placeholder":
-        return (
-          <Sheet
-            variant="outlined"
-            sx={{
-              height: "60px",
-              borderRadius: "md",
-            }}
-          ></Sheet>
-        );
-        case "queue":
-      case "entry":
-        return (
-          <Sheet
-            color={props.current ? "primary" : "neutral"}
-            variant={(props.type == "queue") ? "outlined" : (props.current ? "solid" : "soft")}
-            sx={{
-              height: '60px',
-              borderRadius: "md",
-            }}
-            onMouseOver={() => setCanClose(true)}
-            onMouseLeave={() => setCanClose(false)}
-          >
-            <Stack
-              direction="row"
-              justifyContent="space-between"
-              alignItems="center"
-              spacing={1}
-              sx={{
-                height: "100%",
-                p: 1,
-                pr: 2,
-              }}
-            >
-              <AspectRatio
-                ratio={1}
-                sx={{
-                  flexBasis: "calc(60px - 12px)",
-                  borderRadius: "9px",
-                }}
-              >
-                {image ? (
-                  <img src={image} alt="album art" />
-                ) : (
-                  <CircularProgress size="sm" />
-                )}
-              </AspectRatio>
-              <Stack direction="row" sx={{ flexGrow: 1 }} spacing={1}>
-                <Stack direction="column" sx={{ width: "calc(25%)" }}>
-                  <Typography level="body4" sx={{ mb: -1 }}>
-                    SONG
-                  </Typography>
-                  <Typography
-                    sx={{
-                      whiteSpace: "nowrap",
-                      overflow: "hidden",
-                      textOverflow: "ellipsis",
-                    }}
-                  >
-                    {props.releaseTitle}
-                  </Typography>
-                </Stack>
-                <Stack direction="column" sx={{ width: "calc(25%)" }}>
-                  <Typography level="body4" sx={{ mb: -1 }}>
-                    ARTIST
-                  </Typography>
-                  <Typography
-                    sx={{
-                      whiteSpace: "nowrap",
-                      overflow: "hidden",
-                      textOverflow: "ellipsis",
-                    }}
-                  >
-                    {props.releaseArtist}
-                  </Typography>
-                </Stack>
-                <Stack direction="column" sx={{ width: "calc(25%)" }}>
-                  <Typography level="body4" sx={{ mb: -1 }}>
-                    ALBUM
-                  </Typography>
-                  <Typography
-                    sx={{
-                      whiteSpace: "nowrap",
-                      overflow: "hidden",
-                      textOverflow: "ellipsis",
-                    }}
-                  >
-                    {props.releaseAlbum}
-                  </Typography>
-                </Stack>
-                <Stack direction="column" sx={{ width: "calc(25%)" }}>
-                  <Typography level="body4" sx={{ mb: -1 }}>
-                    LABEL
-                  </Typography>
-                  <Typography
-                    sx={{
-                      whiteSpace: "nowrap",
-                      overflow: "hidden",
-                      textOverflow: "ellipsis",
-                    }}
-                  >
-                    {props.releaseLabel}
-                  </Typography>
-                </Stack>
-              </Stack>
-              {(canClose && !props.current && props.type == "queue") && (
-                <IconButton
-                    size="sm"
-                    variant="solid"
-                    sx = {{
-                        position: 'absolute',
-                    }}
-                >
-                    <PlayArrowIcon />
-                </IconButton>
-              )}
-              {props.current ? (
-                <IconButton color="neutral" variant="plain" size="sm">
-                  <KeyboardArrowDownIcon />
-                </IconButton>
-              ) : (
-                <IconButton
-                  color="neutral"
-                  variant="plain"
-                  size="sm"
-                  sx={{
-                    cursor: "grab",
-                    "&:hover": {
-                      background: "none",
-                    },
-                  }}
-                >
-                  <DragIndicatorIcon />
-                </IconButton>
-              )}
-            </Stack>
-            {(canClose && !props.current) && (
-            <Button
-                color="neutral"
-                variant="solid"
-                sx = {{
-                    position: 'absolute',
-                    zIndex: 4,
-                    top: '50%',
-                    transform: 'translateY(-50%)',
-                    right: 10,
-                    minWidth: '3px',
-                    minHeight: '3px',
-                    maxWidth: '3px',
-                    maxHeight: '3px',
-                    p: 0,
-                    '& svg': {
-                        width: '15px',
-                        height: '15px',
-                    }
-                }}
-            >
-                <ClearIcon />
-            </Button>)}
-          </Sheet>
-        );
-      case "joined":
-      case "left":
-        return (
-          <Sheet
-            color="info"
-            variant="solid"
-            sx={{
-              height: "40px",
-              borderRadius: "md",
-            }}
-          >
-            <Stack
-              direction="row"
-              justifyContent="space-between"
-              alignItems="center"
-              spacing={1}
-              sx={{
-                height: "100%",
-                p: 1,
-              }}
-            >
-              <Typography color="info">
-                {props.type === "joined" ? (
-                  <HeadphonesIcon sx={{ mb: -0.5 }} />
-                ) : (
-                  <LogoutIcon sx={{ mb: -0.5 }} />
-                )}
-              </Typography>
-              <Typography
-                level="body1"
-                endDecorator={
-                  <Typography color="info">{`${props.type} the set!`}</Typography>
-                }
-              >
-                {props.message?.split(` ${props.type}`)?.[0] ??
-                  "Processing Error"}
-              </Typography>
-              <div></div>
-            </Stack>
-          </Sheet>
-        );
-      case "breakpoint":
-        return (
-          <Sheet
-            color="warning"
-            variant="plain"
-            sx={{
-              height: "40px",
-              borderRadius: "md",
-            }}
-          >
-            <Stack
-              direction="row"
-              justifyContent="space-between"
-              alignItems="center"
-              spacing={1}
-              sx={{
-                height: "100%",
-                p: 1,
-              }}
-            >
-              <Typography color="warning">
-                <TimerIcon sx={{ mb: -0.5 }} />
-              </Typography>
-              <Typography level="body1" color="warning">
-                {props.message ?? "Processing Error"}
-              </Typography>
-              <div></div>
-            </Stack>
-          </Sheet>
-        );
-      case "talkset":
-        return (
-          <Sheet
-            color="success"
-            variant="solid"
-            sx={{
-              height: "40px",
-              borderRadius: "md",
-            }}
-          >
-            <Stack
-              direction="row"
-              justifyContent="space-between"
-              alignItems="center"
-              spacing={1}
-              sx={{
-                height: "100%",
-                p: 1,
-              }}
-            >
-              <Typography
-                color="success"
-                sx={{
-                  alignSelf: "flex-start",
-                }}
-              >
-                <MicIcon sx={{ mb: -0.7 }} />
-              </Typography>
-              <Typography level="body1">
-                {props.message ?? "Processing Error"}
-              </Typography>
-              <div></div>
-            </Stack>
-          </Sheet>
-        );
-      default:
-        return (
-          <Sheet
-            color="danger"
-            variant="solid"
-            sx={{
-              height: "40px",
-              borderRadius: "md",
-            }}
-          >
-            <Stack
-              direction="row"
-              justifyContent="center"
-              alignItems="center"
-              spacing={1}
-              sx={{
-                height: "100%",
-                p: 1,
-              }}
-            >
-              <Typography level="body1">
-                {props.message ?? "Processing Error"}
-              </Typography>
-            </Stack>
-          </Sheet>
-        );
-    }
   };
   
   export default FlowSheetPage;
